Improved constructor method
Now the constructor method it has an Enum argument to determine if the Optimization that will use the object is with actived or inactived events, or without optimization at all. This is because the inactivation of the events is error-prone if the programmer not deals with the inner members of an object in a trustworthy way.